requirements.txt
flask
psycopg2-binary
python-dotenv
flask-jwt-extended
flask_swagger_ui